fix admin functions

+++ Services/TransactionService.py
@@ -2,12 +2,15 @@ from Database.TransactionsDAO import TransactionsDAO as Tdao
 from Models.transaction import transaction
 from datetime import datetime
 from Util.validationError import ValidationError
+from bson import ObjectId
 
 class TransactionService():
     def __init__(self, dao: Tdao):
         self.__dao = dao
 
     def getTransactionById(self, transactionId):
+        if not ObjectId.is_valid(transactionId):
+            raise ValidationError("Improper transaction ID")
         return self.__dao.getTransaction(transactionId)
 
     def getAllTransactions(self):
@@ -19,17 +22,30 @@ class TransactionService():
         except ValueError:
             raise ValidationError("Invalid Date")
         
+        if not ObjectId.is_valid(userID):
+            raise ValidationError("Improper user ID")
+        
         return self.__dao.getUserTransactionByDate(userID, date)
 
     def getUserTransactions(self, userID):
+        if not ObjectId.is_valid(userID):
+            raise ValidationError("Improper user ID")
         return self.__dao.getUserTransactions(userID)
     
     def deleteUserTransactions(self, userID):
+        if not ObjectId.is_valid(userID):
+            raise ValidationError("Improper user ID")
         return self.__dao.deleteUserTransactions(userID)
 
     def createTransaction(self, userID, purchased, amount, description):
         now = datetime.now()
         date_now = now.strftime("%m-%d-%y")
+
+        if not ObjectId.is_valid(userID):
+            raise ValidationError("Improper user ID")
+
+        if type(amount) != int:
+            raise ValidationError('Amount must be a number')
 
         if amount < 0:
             raise ValidationError('Amount must be greater than 0')
@@ -39,6 +55,13 @@ class TransactionService():
         return self.__dao.createTransaction(new_transaction)
     
     def updateTransaction(self, transactionID, purchased, amount, description):
+
+        if not ObjectId.is_valid(transactionID):
+            raise ValidationError("Improper transaction ID")
+
+        if type(amount) != int:
+            raise ValidationError('Amount must be a number')
+        
         if amount < 0:
             raise ValidationError('Amount must be greater than 0')
         
@@ -47,6 +70,9 @@ class TransactionService():
         return self.__dao.updateTransaction(new_transaction)
     
     def deleteTransaction(self, transactionID):
+        if not ObjectId.is_valid(transactionID):
+            raise ValidationError("Improper transaction ID")
+        
         return self.__dao.deleteTransaction(transactionID)
